Initial Commit
Project was created for Western Governors University (WGU) - Software Development - User Experience Design (UXDesign). \n- The project studied the software development life cycle in a User Experience Centered view. \n- I took the liberty to create and practice HTML/CSS/JS skills to create a fleshed out website. \n- The project follows the ficitional story that the governement of Taniti is planning to redo it's tourist-centered travel website. \n- The initial commit was turned in to WGU for a final project, it recieved a 100% critieria score.